requirements.txt
anyio==3.5.0
asgiref==3.5.0
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.2
fastapi==0.75.2
h11==0.13.0
idna==3.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pydantic==1.9.0
pyparsing==3.0.9
pytest==7.1.2
requests==2.28.1
sniffio==1.2.0
starlette==0.17.1
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.10
uvicorn==0.17.6
